Added functionality for changing params mid sim
Previously, the simulation could only be started and run
once with the initial parameters. I added functionality to
run the simulation, pause it, change the parameters, and then
continue.

I did this by generalizing the calculateData() function to
calculate the array values at the index value at which the
state variable is in. (If the simulation has just started,
the index = 0, therefore it's just the same. But if the simulation
has started, then the index value =/= 0 and the arrays are
computed from the new starting point as given by the slice() method